Restructure files for events
