+++ internal/audit/auditer/auditer.go
@@ -2,6 +2,7 @@ package auditer
 
 import (
 	"context"
+
 	"github.com/nais/api/internal/auditevent"
 	"github.com/nais/api/internal/auth/authz"
 	"github.com/nais/api/internal/database"
@@ -22,4 +23,8 @@ func (a *Auditer) TeamAddMember(ctx context.Context, actor authz.AuthenticatedUs
 
 func (a *Auditer) TeamRemoveMember(ctx context.Context, actor authz.AuthenticatedUser, team slug.Slug, memberEmail string) error {
 	return a.db.CreateAuditEvent(ctx, auditevent.NewTeamRemoveMember(actor, team, memberEmail))
+}
+
+func (a *Auditer) TeamSetMemberRole(ctx context.Context, actor authz.AuthenticatedUser, team slug.Slug, memberEmail string, role string) error {
+	return a.db.CreateAuditEvent(ctx, auditevent.NewTeamSetMemberRole(actor, team, memberEmail, role))
 }
